TESTED dirty slow unsafe single-core

+++ hash_STUB.h
@@ -8,7 +8,7 @@
 int my_hash_function(std::string* plain_text_input, s_byte_array* hash_output)
 {
   
-  hash_output->size = (u32)sizeof(std::size_t);
+  hash_output->size = (u64)sizeof(std::size_t);
   hash_output->value = (u8*)malloc(hash_output->size);
 
   if(hash_output->value == nullptr) return 1;
